Rename string splice
