+++ src/redux/note-details/format-selection/formatters/utils/string-splice.ts
